refactor(server): split code_health router into a package
The code-health router had grown to ~1088 lines: 13 route handlers plus
about 30 serializers, aggregation helpers, score-breakdown logic, badge
rendering, and async DB loaders in one module.

Split it into a code_health/ package where the shared APIRouter lives in
_router.py and the routes are grouped by resource (overview, findings,
files, trends, coverage, refactoring-targets, churn, badge), backed by
leaf helper modules (serializers, aggregation, breakdown, loaders).

Pure structural move: every symbol relocates verbatim, lazy imports stay
function-local, and __init__.py imports the route modules for their
decorator side-effects and re-exports the helpers that routers/files.py
and the tests import, so the public path repowise.server.routers.code_health
still resolves router and all helpers unchanged.
